Make build[1] and test[1] work on Linux.
[1]: Since this project was created using the Swift Package Manager, the
build piece was really easy. It just built, first try.

[2]: Testing, on the other hand, was (and is) a pain in the ass. It's a
well-known issue, but the lack of introspection in the base test case
framework on Linux means that you have to create a LinuxMain.swift file,
which manually imports all the test cases, and the normal way of doing
that requires manually implementing an allTests computed property, for
every single test case class in the project, and manually updating the
return value of that property every time you add a new test method (or
change the name). Unfortunate, but that's the current status quo.