Simplify DerpSerializationValues so that it works on Linux.
The (in-development, unfinished) Foundation library for Linux is
behind macOS; NSJSONSerialization is still flakey. But, by
changing DerpSerializationValues to disallow nil, ( [String:Any]
instead of [String: Any?] ), I was able to get it working on:

swift-3.1-DEVELOPMENT-SNAPSHOT-2017-02-11-a-ubuntu16.04

This required a few changes where support for nil was being used
by objects adopting DerpSerializable.